:art: Updated structure of code
