+++ conanfile.txt
@@ -1,5 +1,9 @@
 [requires]
 libhal-lpc40xx/0.3.10
+libhal-rmd/0.3.6
+libhal-esp8266/0.3.4
+libhal-util/0.3.8
+
 
 [tool_requires]
 gnu-arm-embedded-toolchain/11.3.0
