+++ libhal-bash-scripts/.README.md
@@ -1,0 +1,38 @@
+NOTE: This is how you setup a project for the first time after installing 
+When you reach "Cloning the target libraries" in libhal do the following:
+    run:
+        git clone https://github.com/libhal/liblpc40xx
+        cd liblpc40xx/demos
+        mkdir build
+        cd build
+        conan install .. -s build_type=Debug
+Then depending on your operating system do the following:
+    WSL:
+        cd ..
+        run ./build_debug_wsl.sh
+    Windows:
+        cd ..
+        run ./build_debug_windows.sh
+    Linux:
+        cd ..
+        run chmod +x build_debug_wsl.sh
+    MacOS:
+        TBD
+The previous will run a bash script which will build and compile your project for you and if it doesn't work, then open the .sh
+file and run each command on your own.
+
+Flashing:
+    This is relatively simple, do the following:
+        If you haven't downloaded nxpprog already run: 
+            pip3 install nxpprog
+        Then use the following command but goto the build file and run ls, then look for the binary file you are trying to flash, replace lpc4078_uart.bin with the binary file name then goto https://libhal.github.io/web-serial/ and click connect when you have your dev-2 board plugged in, here you will be able to click connect, scroll all the way down and find what COM0X port the de-2 is connected to, switch the /dev/tty.usbserial-140 in the following command to that port and run the commmand:
+            nxpprog --control --binary="lpc4078_app.bin " --device="COM5"
+        
+After this you are all set to continue flashing and building using the previous instructions.
+If you are in windows and see a lot of printouts with a final statement of 'force_update', this is expected an your code still flashed it is a current bug being worked on.
+
+Use the following link to find the different libaries that are supported:
+    https://libhal.jfrog.io/ui/packages?name=lib&type=packages
+
+If you want to include another library in the project you copy and past a library name then / then the version number into the conanfile.txt like the following for the rmd library (put it under requires):
+    librmd/0.3.2
