+++ libhal-bash-scripts/build_debug.sh
