+++ libhal-bash-scripts/build_release.sh
