+++ libhal.tweaks.hpp
